Rename outline-table-view to outline-view-table
Correcting initial misnamed package to OutlineView base naming convention (OutlineViewTable, OutlineViewList, etc)
