--- packages/terra-outline-table-view/src/Table.module.scss
@@ -1,11 +1,0 @@
-// Themes
-@import './clinical-lowlight-theme/Table.module';
-@import './orion-fusion-theme/Table.module';
-
-:local {
-  // Standard block display
-  // ----------------------
-  .outline {
-    border: 1px solid #dedfe0;
-  }
-}
