--- packages/terra-outline-table-view/src/index.js
@@ -1,3 +1,0 @@
-import Table from './Table';
-
-export default Table;
