--- packages/terra-outline-table-view/src/orion-fusion-theme/Table.module.scss
@@ -1,7 +1,0 @@
-@import '~terra-mixins/lib/Mixins';
-
-:local {
-  .outline {
-    border: 1px solid #dedfe0;
-  }
-}
