--- packages/terra-outline-table-view/src/terra-dev-site/doc/outline-table-view/ChangeLog.2.doc.mdx
@@ -1,6 +1,0 @@
-import { Badge } from 'terra-outline-table-view/package.json?dev-site-package';
-import ChangeLog from 'terra-outline-table-view/CHANGELOG.md';
-
-<Badge />
-
-<ChangeLog />
