--- packages/terra-outline-table-view/tests/wdio/table-spec.js
@@ -1,6 +1,0 @@
-Terra.describeViewports('Table', ['medium'], () => {
-  it('displays outline table view', () => {
-    browser.url('/raw/tests/terra-outline-table-view/outline-table-view/outline-table-view');
-    Terra.validates.element('default');
-  });
-});
