+++ packages/terra-outline-view-table/CHANGELOG.md
@@ -3,10 +3,10 @@
 ## Unreleased
 
 * Added
-  * Added new outline-table-view docs and tests folders
+  * Added new outline-view-table docs and tests folders
 
 * Changed
-  * terra-outline-table-view tests and doc for now initial implementation
+  * terra-outline-view-table tests and doc for now initial implementation
 
 * Added
   * Remove before release: Set base component props for pass through with a couple exceptions
