+++ packages/terra-outline-view-table/src/clinical-lowlight-theme/OutlineViewTable.module.scss
@@ -1,0 +1,6 @@
+:local {
+  .clinical-lowlight-theme {
+    // TODO: Replace with actual styles
+    --terra-outline-view-table-foo-bar: none;
+  }
+}
