+++ packages/terra-outline-view-table/src/orion-fusion-theme/OutlineViewTable.module.scss
@@ -1,0 +1,6 @@
+:local {
+  .orion-fusion-theme {
+    // TODO: Replace with actual styles
+    --terra-outline-view-table-foo-bar: none;
+  }
+}
