+++ packages/terra-outline-view-table/src/proptypes/cellShape.js
