+++ packages/terra-outline-view-table/src/proptypes/discloseShape.js
