+++ packages/terra-outline-view-table/src/proptypes/headerCellShape.js
