+++ packages/terra-outline-view-table/src/proptypes/headerShape.js
