+++ packages/terra-outline-view-table/src/proptypes/rowShape.js
