+++ packages/terra-outline-view-table/src/proptypes/sectionHeaderShape.js
