+++ packages/terra-outline-view-table/src/proptypes/sectionShape.js
