+++ packages/terra-outline-view-table/src/proptypes/selectAllColumnShape.js
