+++ packages/terra-outline-view-table/src/proptypes/toggleShape.js
