+++ packages/terra-outline-view-table/src/proptypes/widthShape.js
