+++ packages/terra-outline-view-table/src/terra-dev-site/doc/outline-view-table/About.1.doc.mdx
@@ -1,17 +1,17 @@
-import { Badge } from 'terra-table/package.json?dev-site-package';
+import { Badge } from 'terra-outline-view-table/package.json?dev-site-package';
 import DefaultTable from './example/DefaultTable?dev-site-example';
-import PropsTable from 'terra-outline-table-view/src/Table?dev-site-props-table';
+import PropsTable from 'terra-outline-view-table/src/OutlineViewTable?dev-site-props-table';
 
 <Badge />
 
-# Terra Outline Table View
+# Terra Outline View Table
 
-Terra Outline Table View is a structural component used to
+Terra Outline View Table is a structural component used to ... _(add content)_ <!-- TODO: add more content -->
 
 ## Getting Started
 
 - Install with [npmjs](https://www.npmjs.com):
-  - `npm install terra-outline-table-view`
+  - `npm install terra-outline-view-table`
 
 <!-- AUTO-GENERATED-CONTENT:START Peer Dependencies -->
 ## Peer Dependencies
@@ -29,7 +29,7 @@ This component requires the following peer dependencies be installed in your app
 ## Usage
 
 ```jsx
-import Table from 'terra-outline-table-view';
+import OutlineViewTable from 'terra-outline-view-table';
 ```
 
 ## Component Features
@@ -41,8 +41,8 @@ import Table from 'terra-outline-table-view';
 
 # Examples
 
-<DefaultTable />
+<DefaultTable title="Sample" description="Sample passthrough props to table example. Replace with an outline-view-table example when ready." />
 
-## Table Props
+## Outline View Table Props
 
 <PropsTable />
