+++ packages/terra-outline-view-table/src/terra-dev-site/doc/outline-view-table/ChangeLog.2.doc.mdx
@@ -1,0 +1,6 @@
+import { Badge } from 'terra-outline-view-table/package.json?dev-site-package';
+import ChangeLog from 'terra-outline-view-table/CHANGELOG.md';
+
+<Badge />
+
+<ChangeLog />
