+++ packages/terra-outline-view-table/src/terra-dev-site/test/outline-view-table/OutlineViewTable.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import OutlineTableView from 'terra-outline-table-view';
+import OutlineTableView from 'terra-outline-view-table';
 
 const DefaultTest = () => (
   <OutlineTableView
